feat: create cration modal
