+++ src/routes/phone-search/+page.svelte
@@ -1,10 +1,14 @@
 <script>
 	import { Modal, SimpleGrid, Card, Text, Button, Input, Flex, TextInput } from '@svelteuidev/core';
+	import { CONTACT_MODAL, formatNumber } from '$lib/utils/contacts-utils.js'
+	import ContactInfo from '$lib/components/contacts/contact-info.svelte'
+	
 	import MdiPen from 'virtual:icons/mdi/pen';
 
 	export let data;
 	let number = '';
 	let showModal = false;
+	let showCreationModal = false;
 	let modalData = {};
 
 	const handleModalOpen = () => {
@@ -23,37 +27,17 @@
 		showModal = false;
 	};
 
-	// all the logic below is to format the phone number like this: (214) 180-1323
-	// since it is saved like this in airtable and we need to filter and search for it in the exact fashion
-	// lemme know if you have a smarter way to do this.
-	const formatNumber = (number) => {
-		const individualNumbers = number.split('').filter((n) => !isNaN(n) && n !== ' ');
-		let formattedNumber = '';
-		individualNumbers.forEach((num, i) => {
-			if (i === 0) {
-				formattedNumber = `(${num}`;
-			} else if (i === 2) {
-				formattedNumber += `${num})`;
-			} else if (i == 3) {
-				formattedNumber += ` ${num}`;
-			} else if (i === 5) {
-				formattedNumber += `${num}-`;
-			} else {
-				formattedNumber += num;
-			}
-		});
-		return formattedNumber;
-	};
-
 	const handleNumber = (e) => {
 		let num = e.target.value;
 		number = formatNumber(num);
 	};
 
-	const checkNumberValidity = (number) => {
-		const formattedNumber = formatNumber(number);
-		modalData['Phone Number'] = formattedNumber;
-	};
+	const handleOpenCreateModal = (e) => {
+		e.preventDefault();
+		modalData = { ...CONTACT_MODAL}
+		showCreationModal = true;
+	}
+
 </script>
 
 <main>
@@ -68,7 +52,7 @@
 				placeholder="Enter phone number"
 			/>
 			<Button type="submit">Search</Button>
-			<Button>Create</Button>
+			<Button on:click={handleOpenCreateModal}>Create</Button>
 		</Flex>
 	</form>
 	<SimpleGrid style="margin: 1em" cols={2}>
@@ -100,33 +84,19 @@
 	<Modal opened={showModal} on:close={closeModal} title="Details">
 		<form method="POST" action="?/updateRecord">
 			<input name="id" type="hidden" value={data.records[0].id} />
-			<TextInput
-				required
-				label="First Name"
-				name="First Name"
-				bind:value={modalData['First Name']}
-			/>
-			<TextInput required label="Last Name" name="Last Name" bind:value={modalData['Last Name']} />
-			<TextInput
-				required
-				label="Phone Number"
-				name="Phone Number"
-				maxlength="14"
-				bind:value={modalData['Phone Number']}
-				on:input={(e) => checkNumberValidity(e.target.value)}
-			/>
-			<TextInput
-				required
-				label="Street Address"
-				name="Street Address"
-				bind:value={modalData['Street Address']}
-			/>
-			<TextInput required label="City" name="City" bind:value={modalData['City']} />
-			<TextInput required label="Zip" name="Zip" bind:value={modalData['Zip']} />
-			<TextInput required label="State" name="State" bind:value={modalData['State']} />
+			<ContactInfo modalData={modalData} />
 			<Flex gap="sm" mt={10} justify="right">
 				<Button variant="outline" on:click={closeModal}>Cancel</Button>
 				<Button type="submit" on:click={() => console.log(data)}>Save</Button>
+			</Flex>
+		</form>
+	</Modal>
+	<Modal opened={showCreationModal} on:close={() => {showCreationModal = false}} title="Create Contact" >
+		<form method="POST" action="?/createRecord">
+			<ContactInfo modalData={modalData}/>
+			<Flex gap="sm" mt={10} justify="right">
+				<Button variant="outline" on:click={closeModal}>Cancel</Button>
+				<Button>Create</Button>
 			</Flex>
 		</form>
 	</Modal>
